feat(LayeredContainers): add layered container with content and glass layers
